+++ end-to-end-tests/script/Transfer.s.sol
@@ -14,15 +14,23 @@ import { Utils } from "@eveworld/world/src/modules/inventory/Utils.sol";
 import { SmartStorageUnitLib } from "@eveworld/world/src/modules/smart-storage-unit/SmartStorageUnitLib.sol";
 import { InventoryLib } from "@eveworld/world/src/modules/inventory/InventoryLib.sol";
 import { FRONTIER_WORLD_DEPLOYMENT_NAMESPACE as DEPLOYMENT_NAMESPACE } from "@eveworld/common-constants/src/constants.sol";
+import { SystemRegistry } from "@latticexyz/world/src/codegen/tables/SystemRegistry.sol";
 
 contract Transfer is Script {
+  // assumes CreateAndAnchor.s.sol, BringOnline.s.sol and CreateAndDeposit.s.sol have been run
   using InventoryLib for InventoryLib.World;
+  using SmartStorageUnitLib for SmartStorageUnitLib.World;
   using Utils for bytes14;
 
   function run(address worldAddress) public {
     StoreSwitch.setStoreAddress(worldAddress);
 
     InventoryLib.World memory inventory = InventoryLib.World({
+      iface: IBaseWorld(worldAddress),
+      namespace: DEPLOYMENT_NAMESPACE
+    });
+
+    SmartStorageUnitLib.World memory SSUInterface = SmartStorageUnitLib.World({
       iface: IBaseWorld(worldAddress),
       namespace: DEPLOYMENT_NAMESPACE
     });
@@ -39,9 +47,8 @@ contract Transfer is Script {
     address ownerSSU = vm.addr(ownerPrivateKey);
 
     // // Start broadcasting transactions from the deployer account
-    vm.startBroadcast(ephemeralPrivateKey);
 
-    // CHOOSE WHICH ITEM TO MOVE FROM INVENTORY TO EPHEMERAL
+    // ITEM TO MOVE FROM INVENTORY TO EPHEMERAL
     uint256 inventoryItemId = uint256(123);
     InventoryItem[] memory invItems = new InventoryItem[](1);
     invItems[0] = InventoryItem({
@@ -53,7 +60,7 @@ contract Transfer is Script {
       quantity: 1
     });
 
-    // CHOOSE WHICH ITEM TO MOVE FROM EPHEMERAL TO INVENTORY
+    // ITEM TO MOVE FROM EPHEMERAL TO INVENTORY
     uint256 ephInventoryItemId = uint256(345);
     InventoryItem[] memory ephInvItems = new InventoryItem[](1);
     ephInvItems[0] = InventoryItem({
@@ -65,46 +72,30 @@ contract Transfer is Script {
       quantity: 1
     });
 
-    //Before transfer
-    //SSU owner should have 0 ephInvItem
-    //Ephermeral owner should have 0 invItem
-    InventoryItemTableData memory ephInvItemData = InventoryItemTable.get(
-      DEPLOYMENT_NAMESPACE.inventoryItemTableId(),
-      smartObjectId,
-      ephInvItems[0].inventoryItemId
-    );
-    console.log(ephInvItemData.quantity); //0
-
-    EphemeralInvItemTableData memory invItem = EphemeralInvItemTable.get(
-      DEPLOYMENT_NAMESPACE.ephemeralInventoryItemTableId(),
-      smartObjectId,
-      invItems[0].inventoryItemId,
-      ephemeralOwner
-    );
-    console.log(invItem.quantity); //0
+    vm.startBroadcast(ephemeralPrivateKey); // if triggered the createItem functions need to be called from the owner/deployer account
 
     // TRANSFER
     inventory.inventoryToEphemeralTransfer(smartObjectId, invItems);
     inventory.ephemeralToInventoryTransfer(smartObjectId, ephInvItems);
 
-    //After trasnfer 1 invItem should go into ephemeral and 1 ephInvItem should go into inventory
-    //SSU owner should have 1 ephInvItem after transfer
-    //Ephermeral owner should have 1 invItem after transfer
+    // After transfer 1 invItem should go into ephemeral and 1 ephInvItem should go into inventory
+    // SSU owner should have 1 ephInvItem after transfer
+    // Ephermeral owner should have 1 invItem after transfer
 
-    ephInvItemData = InventoryItemTable.get(
+    InventoryItemTableData memory ephItemInInv = InventoryItemTable.get(
       DEPLOYMENT_NAMESPACE.inventoryItemTableId(),
       smartObjectId,
       ephInvItems[0].inventoryItemId
     );
-    console.log(ephInvItemData.quantity); //1
+    console.log(ephItemInInv.quantity); //1
 
-    invItem = EphemeralInvItemTable.get(
+    EphemeralInvItemTableData memory itemInEphInv = EphemeralInvItemTable.get(
       DEPLOYMENT_NAMESPACE.ephemeralInventoryItemTableId(),
       smartObjectId,
       invItems[0].inventoryItemId,
       ephemeralOwner
     );
-    console.log(invItem.quantity); //1
+    console.log(itemInEphInv.quantity); //1
 
     // STOP THE BROADCAST
     vm.stopBroadcast();